// src/pharmacy/types/enums/pharmacy.enum.ts
export enum PharmacyRejectReasons {
  MEDICINES_NOT_AVAILABLE = 'MEDICINES_NOT_AVAILABLE',
  PRESCRIPTION_NOT_VALID = 'PRESCRIPTION_NOT_VALID',
  PRESCRIPTION_NOT_CLEAR = 'PRESCRIPTION_NOT_CLEAR',
  EMIRATES_ID_NOT_CORRECT = 'EMIRATES_ID_NOT_CORRECT'
}

export enum PharmacyAcceptIssues {
  HEALTH_CARD_NOT_VALID = 'HEALTH_CARD_NOT_VALID',
  HEALTH_CARD_EXPIRED = 'HEALTH_CARD_EXPIRED',
  HEALTH_CARD_NOT_CLEAR = 'HEALTH_CARD_NOT_CLEAR'
}